DatastoreIO: move to GCP module
