+++ examples/java/src/main/java/org/apache/beam/examples/complete/AutoComplete.java
@@ -25,9 +25,9 @@ import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.PipelineResult;
 import org.apache.beam.sdk.coders.AvroCoder;
 import org.apache.beam.sdk.coders.DefaultCoder;
-import org.apache.beam.sdk.io.DatastoreIO;
 import org.apache.beam.sdk.io.TextIO;
 import org.apache.beam.sdk.io.gcp.BigQueryIO;
+import org.apache.beam.sdk.io.gcp.DatastoreIO;
 import org.apache.beam.sdk.io.gcp.PubsubIO;
 import org.apache.beam.sdk.options.Default;
 import org.apache.beam.sdk.options.Description;
