+++ examples/java/src/main/java/org/apache/beam/examples/cookbook/DatastoreWordCount.java
@@ -25,9 +25,9 @@ import static com.google.api.services.datastore.client.DatastoreHelper.makeValue
 
 import org.apache.beam.examples.WordCount;
 import org.apache.beam.sdk.Pipeline;
-import org.apache.beam.sdk.io.DatastoreIO;
 import org.apache.beam.sdk.io.Read;
 import org.apache.beam.sdk.io.TextIO;
+import org.apache.beam.sdk.io.gcp.DatastoreIO;
 import org.apache.beam.sdk.options.Default;
 import org.apache.beam.sdk.options.Description;
 import org.apache.beam.sdk.options.PipelineOptions;
